fix for header loop
