+++ include/aquarius.hpp
@@ -7,6 +7,7 @@
 #include <aquarius/error_code.hpp>
 #include <aquarius/io_service_pool.hpp>
 #include <aquarius/ip/handler.hpp>
+#include <aquarius/ip/handler_reg.hpp>
 #include <aquarius/ip/http.hpp>
 #include <aquarius/ip/tcp.hpp>
 #include <aquarius/ip/udp.hpp>
