+++ include/aquarius/basic_context.hpp
@@ -1,5 +1,4 @@
 #pragma once
-#include <aquarius/detail/asio.hpp>
 #include <aquarius/detail/flex_buffer.hpp>
 #include <aquarius/error_code.hpp>
 #include <aquarius/ip/handler_error.hpp>
